+++ src/Components/Header/Header.jsx
@@ -1,9 +1,33 @@
 import LogoDigital from "../../assets/Logo-Digital.png"
+import { useState, useEffect } from "react"
 import IconCartShop from '../../assets/Icon-cartShop.png'
+import usericon from '../../assets/usericon.svg'
 import './Header.css'
 import { Link } from 'react-router-dom'
 
 function Header() {
+
+    const [user, setUser] = useState()
+
+    useEffect(() => {
+        // Recupera o usuário do localStorage
+        const storedUser = localStorage.getItem('user');
+        if (storedUser) {
+            setUser(JSON.parse(storedUser));
+        }
+    }, []);
+
+    let conteudo 
+    if (user){
+        conteudo =   <Link className="acc-infos" to="/meuspedidos"><img src={usericon} alt="" />Olá {user.firstname}</Link>
+    } else{
+        conteudo = (
+            <>
+                <Link to="/criarLogin"><div id='Register'><a href="#">Cadastre-se</a></div></Link>
+                <Link to="/login"><button id='Btn-Enter'><a href="#">Entrar</a></button></Link>
+            </>
+        )
+    }
 
     return(
         <>
@@ -15,8 +39,7 @@ function Header() {
                     <input type="text" placeholder='Pesquisar produto...'/>
                 </div>
                 <div id='Access-Part'>
-                    <Link to="/criarLogin"><div id='Register'><a href="#">Cadastre-se</a></div></Link>
-                    <Link to="/login"><button id='Btn-Enter'><a href="#">Entrar</a></button></Link>
+                    {conteudo}
                     <div id='Icon-CartShop'><img src={IconCartShop}  alt="icon-de-compra" /></div>
                 </div>
             </header>
